Fix pointer-to-int and int-to-pointer cast warnings in demo programs
Use intermediate long cast to avoid direct pointer-integer conversions:

exm_in_c.c:
- (XtPointer)menu_num -> (XtPointer)(long)menu_num
- (int)cd -> (int)(long)cd (2 instances)

todo.c:
- (int) item -> (int)(long) item

tooltips.c:
- (int)client_data -> (int)(long)client_data

Fixes clang warnings about casting between pointer and integer types
Ensures proper type safety on 64-bit architectures where pointers are larger than int

+++ demos/programs/Exm/app_in_c/exm_in_c.c
@@ -235,7 +235,7 @@ CreateMenus(Widget parent_of_menu_bar)
 				XmNmnemonic, mnemonic_char,
                                 NULL);
       XtAddCallback (MenuButtonWidget[menu_num], XmNactivateCallback,
-                     HelpCB, (XtPointer)menu_num);
+                     HelpCB, (XtPointer)(long)menu_num);
       XmStringFree(menu_as_a_cs);
    }
 
@@ -469,7 +469,7 @@ HelpCB(Widget w,
        XtPointer cd,
        XtPointer cb)
 {
- int       what_kind_of_help = (int)cd;
+ int       what_kind_of_help = (int)(long)cd;
 static char      *messages[] = {
 "ExmSimple displays one oval or rectangle.\n",
 "ExmString displays one compound string.\n",
@@ -500,7 +500,7 @@ as a button inside a menu.",
  Widget    help_dialog;
  Arg       arg[3];
 
-   message_as_a_cs = XmStringLtoRCreate(messages[(int)cd],
+   message_as_a_cs = XmStringLtoRCreate(messages[(int)(long)cd],
                                         XmFONTLIST_DEFAULT_TAG);
 
    XtSetArg(arg[0], XmNmessageString, message_as_a_cs);
